+++ CHANGELOG.md
@@ -2,6 +2,14 @@ The upgrade instructions are available at [Oro documentation website](https://do
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
 
+## UNRELEASED
+
+### Changed
+
+#### ProductBundle
+* Storefront product autocomplete now includes list of categories with found products  
+* ProcessAutocompleteDataEvent data format has been changed, now it includes full autocomplete data: products, categories, and total count
+
 ## 5.1.0-beta.2 (2022-11-30)
 [Show detailed list of changes](incompatibilities-5-1-beta-2.md)